Feat: upgrade `MemberReputation` component
