+++ src/components/shared/MemberReputation/index.ts
@@ -1,2 +1,1 @@
 export { default } from './MemberReputation';
-export { default as MemberReputationUI } from './MemberReputationUI';
